requirements.txt
flask==3.1.0
google-generativeai>=0.8.4
requests==2.32.3
python-dotenv==1.1.0
gunicorn==21.2.0 